// model/asyncTaskResponse.ts
/**
 * Megaputer Text Mining API
 * Megaputer Text Mining API
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export class AsyncTaskResponse {
    /**
    * Identifier of the newly created task, which has been started on the server.
    */
    'taskId'?: string;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "taskId",
            "baseName": "taskId",
            "type": "string"
        }    ];

    static getAttributeTypeMap() {
        return AsyncTaskResponse.attributeTypeMap;
    }
}