// model/entitiesResponse.ts
/**
 * Megaputer Text Mining API
 * Megaputer Text Mining API
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { EntitiesResponseDocuments } from './entitiesResponseDocuments';

export class EntitiesResponse {
    'documents'?: Array<EntitiesResponseDocuments>;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "documents",
            "baseName": "documents",
            "type": "Array<EntitiesResponseDocuments>"
        }    ];

    static getAttributeTypeMap() {
        return EntitiesResponse.attributeTypeMap;
    }
}