// model/entitiesResponseDocuments.ts
/**
 * Megaputer Text Mining API
 * Megaputer Text Mining API
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { EntitiesResponseEntities } from './entitiesResponseEntities';

export class EntitiesResponseDocuments {
    /**
    * Index of document in input documents array
    */
    'id'?: number;
    /**
    * Result of entities operation
    */
    'entities'?: Array<EntitiesResponseEntities>;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "id",
            "baseName": "id",
            "type": "number"
        },
        {
            "name": "entities",
            "baseName": "entities",
            "type": "Array<EntitiesResponseEntities>"
        }    ];

    static getAttributeTypeMap() {
        return EntitiesResponseDocuments.attributeTypeMap;
    }
}